requirements.txt
cirq
cirq-ionq
ionq
python-dotenv
cryptography